+++ src/modules/rules/index.ts
@@ -9,6 +9,7 @@ import UpdatePublicationTitle from "./field-publication";
 import TitleGuillemet from "./field-title-guillemet";
 import TitleSentenceCase from "./field-title-sentence-case";
 import NoDuplicatItem from "./no-duplicate-item";
+import NoExtraZeros from "./no-extra-zeros";
 import NoWebPageItem from "./no-webpage-item";
 import UpdateMetadata from "./retrive-metadata";
 import SetFieldValue from "./set-field-value";
@@ -26,5 +27,6 @@ export {
     TitleGuillemet,
     SetFieldValue,
     NoDuplicatItem,
+    NoExtraZeros,
     NoWebPageItem,
 };